refactored additions to FsUtils into ALAFsUtils
