+++ pipelines/pom.xml
@@ -162,7 +162,7 @@
 
     <!-- GBIF libraries -->
     <gbif-parsers.version>0.46</gbif-parsers.version>
-    <gbif-pipelines.version>2.4.4-SNAPSHOT</gbif-pipelines.version>
+    <gbif-pipelines.version>2.4.8-SNAPSHOT</gbif-pipelines.version>
     <dwca-io.version>2.8</dwca-io.version>
     <gbif-api.version>0.98</gbif-api.version>
     <gbif-common.version>0.47</gbif-common.version>
