+++ pipelines/src/main/java/au/org/ala/kvs/cache/GeocodeServiceFactory.java
@@ -6,18 +6,18 @@ import org.gbif.kvs.cache.KeyValueCache;
 import org.gbif.kvs.geocode.LatLng;
 import org.gbif.pipelines.kv.BitmapFactory;
 import org.gbif.pipelines.parsers.config.model.KvConfig;
-import org.gbif.pipelines.parsers.parsers.location.GeocodeService;
+import org.gbif.pipelines.parsers.parsers.location.GeocodeKvStore;
 import org.gbif.rest.client.geocode.GeocodeResponse;
 
 import lombok.SneakyThrows;
 
 /**
- * Factory to get singleton instance of {@link GeocodeService}
+ * Factory to get singleton instance of {@link GeocodeKvStore}
  */
 public class GeocodeServiceFactory {
 
-  public static GeocodeService create(KvConfig config) {
-    return GeocodeService.create(createKvStore(), BitmapFactory.getInstance(config));
+  public static GeocodeKvStore create(KvConfig config) {
+    return GeocodeKvStore.create(createKvStore(), BitmapFactory.getInstance(config));
   }
 
   private static KeyValueStore<LatLng, GeocodeResponse> createKvStore() {
