+++ pipelines/src/main/java/au/org/ala/pipelines/beam/ALASamplingToAvroPipeline.java
@@ -19,11 +19,9 @@ import org.apache.beam.sdk.values.PCollection;
 import org.apache.beam.sdk.values.TupleTag;
 import org.apache.commons.lang.StringUtils;
 import org.apache.hadoop.fs.FileSystem;
-import org.apache.hadoop.fs.Path;
 import org.gbif.pipelines.common.PipelinesVariables;
 import org.gbif.pipelines.ingest.options.InterpretationPipelineOptions;
 import org.gbif.pipelines.ingest.options.PipelinesOptionsFactory;
-import org.gbif.pipelines.ingest.utils.FsUtils;
 import org.gbif.pipelines.io.avro.AustraliaSpatialRecord;
 import org.gbif.pipelines.io.avro.LocationRecord;
 import org.gbif.pipelines.transforms.specific.AustraliaSpatialTransform;
@@ -60,7 +58,7 @@ public class ALASamplingToAvroPipeline {
         String outputPath = ALAFsUtils.buildPathSamplingOutputUsingTargetPath(options);
         log.info("Outputting results to " + outputPath);
 
-        FileSystem fs = FsUtils.getFileSystem(options.getHdfsSiteConfig(), options.getCoreSiteConfig(), "/");
+        FileSystem fs = ALAFsUtils.getFileSystem(options.getHdfsSiteConfig(), options.getCoreSiteConfig(), "/");
 
 
         // Read column headers
@@ -163,15 +161,15 @@ public class ALASamplingToAvroPipeline {
 
         try {
             //obtain column header
-            if (FsUtils.exists(fs, samplingPath)) {
+            if (ALAFsUtils.exists(fs, samplingPath)) {
 
-                Collection<String> samplingFiles = FsUtils.listPaths(fs, samplingPath);
+                Collection<String> samplingFiles = ALAFsUtils.listPaths(fs, samplingPath);
 
                 if (!samplingFiles.isEmpty()) {
 
                     //read the first line of the first sampling file
                     String samplingFilePath = samplingFiles.iterator().next();
-                    String columnHeaderString = new BufferedReader(new InputStreamReader(FsUtils.openInputStream(fs, samplingFilePath))).readLine();
+                    String columnHeaderString = new BufferedReader(new InputStreamReader(ALAFsUtils.openInputStream(fs, samplingFilePath))).readLine();
                     return columnHeaderString.split(",");
 
                 } else {
