+++ pipelines/src/main/java/au/org/ala/pipelines/beam/ALAUUIDMintingPipeline.java
@@ -5,6 +5,7 @@ import au.org.ala.kvs.ALAKvConfigFactory;
 import au.org.ala.kvs.cache.ALAAttributionKVStoreFactory;
 import au.org.ala.kvs.client.ALACollectoryMetadata;
 import au.org.ala.pipelines.common.ALARecordTypes;
+import au.org.ala.utils.ALAFsUtils;
 import lombok.AccessLevel;
 import lombok.NoArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -23,7 +24,6 @@ import org.apache.beam.sdk.values.TypeDescriptor;
 import org.apache.commons.lang.StringUtils;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
-import org.codehaus.plexus.util.FileUtils;
 import org.gbif.dwc.terms.DwcTerm;
 import org.gbif.dwc.terms.Term;
 import org.gbif.dwc.terms.UnknownTerm;
@@ -36,7 +36,6 @@ import org.gbif.pipelines.io.avro.*;
 import org.gbif.pipelines.parsers.utils.ModelUtils;
 import org.gbif.rest.client.configuration.ClientConfiguration;
 
-import java.io.File;
 import java.util.*;
 
 /**
@@ -81,7 +80,7 @@ public class ALAUUIDMintingPipeline {
     public static void run(InterpretationPipelineOptions options) throws Exception {
 
         Pipeline p = Pipeline.create(options);
-        Properties properties = FsUtils.readPropertiesFile(options.getHdfsSiteConfig(), options.getCoreSiteConfig(), options.getProperties());
+        Properties properties = ALAFsUtils.readPropertiesFile(options.getHdfsSiteConfig(), options.getCoreSiteConfig(), options.getProperties());
 
         //build the directory path for existing identifiers
         String alaRecordDirectoryPath = options.getTargetPath() + "/" + options.getDatasetId().trim() + "/1/identifiers/" + ALARecordTypes.ALA_UUID.name().toLowerCase();
@@ -132,7 +131,7 @@ public class ALAUUIDMintingPipeline {
 
         log.info("Transform 2: ALAUUIDRecord ur ->  <uniqueKey, uuid> (assume incomplete)");
 
-        FileSystem fs = FsUtils.getFileSystem(options.getHdfsSiteConfig(), options.getCoreSiteConfig(), null);
+        FileSystem fs = ALAFsUtils.getFileSystem(options.getHdfsSiteConfig(), options.getCoreSiteConfig(), null);
         Path path = new Path(alaRecordDirectoryPath);
 
         if (fs.exists(path)){
