+++ pipelines/src/test/java/au/org/ala/kvs/GeocodeServiceTest.java
@@ -5,7 +5,7 @@ import org.gbif.kvs.cache.KeyValueCache;
 import org.gbif.kvs.geocode.LatLng;
 import org.gbif.pipelines.parsers.config.factory.KvConfigFactory;
 import org.gbif.pipelines.parsers.config.model.KvConfig;
-import org.gbif.pipelines.parsers.parsers.location.GeocodeService;
+import org.gbif.pipelines.parsers.parsers.location.GeocodeKvStore;
 import org.gbif.rest.client.geocode.GeocodeResponse;
 import org.junit.Test;
 
@@ -25,7 +25,7 @@ public class GeocodeServiceTest {
         p.load(new FileInputStream(new File("src/test/resources/pipelines.properties")));
 
         KvConfig kvConfig = KvConfigFactory.create(p, KvConfigFactory.GEOCODE_PREFIX);
-        GeocodeService geoService = GeocodeServiceFactory.create(kvConfig);
+        GeocodeKvStore geoService = GeocodeServiceFactory.create(kvConfig);
         GeocodeResponse resp = geoService.get(LatLng.builder().withLongitude(146.2).withLatitude(-27.9).build());
         assert !resp.getLocations().isEmpty();
         assert resp.getLocations().iterator().next().getCountryName().equals("AU");
@@ -40,7 +40,7 @@ public class GeocodeServiceTest {
         Properties p = new Properties();
         p.load(new FileInputStream(new File("src/test/resources/pipelines.properties")));
         KvConfig kvConfig = KvConfigFactory.create(p, KvConfigFactory.GEOCODE_PREFIX);
-        GeocodeService geoService = GeocodeServiceFactory.create(kvConfig);
+        GeocodeKvStore geoService = GeocodeServiceFactory.create(kvConfig);
         GeocodeResponse resp = geoService.get(LatLng.builder().withLongitude(151.329751).withLatitude(-36.407357).build());
         assert !resp.getLocations().isEmpty();
         assert resp.getLocations().iterator().next().getCountryName().equals("AU");
